requirements.txt
alembic==1.15.2
altgraph==0.17.2
astroid==3.3.10
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.9.1
cryptography==44.0.2
cssbeautifier==1.15.4
dill==0.4.0
dnspython==2.7.0
EditorConfig==0.17.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
flake8==7.3.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-MySQL==1.6.0
flask-mysql-connector==1.1.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
future==0.18.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.0
macholib==1.15.2
Mako==1.3.9
mando==0.7.1
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.250527
pathspec==0.12.1
pip-chill==1.0.3
platformdirs==4.3.8
pluggy==1.5.0
pycodestyle==2.14.0
pycparser==2.22
pydeps==3.0.1
pyflakes==3.4.0
pylint==3.3.7
PyMySQL==1.1.1
pytest==8.3.5
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
radon==6.0.1
regex==2024.11.6
requests==2.32.3
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.40
stdlib-list==0.11.1
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
types-PyMySQL==1.1.0.20250516
types-pytz==2025.2.0.20250516
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
